refactor(layout): add Layout class and enums for canvas fit and alignment
The Layout class provides methods to manage the fit and alignment of a Rive animation within an HTML canvas. It includes the following features:

- Fit enum with options: Cover, Contain, Fill, FitWidth, FitHeight, None, ScaleDown
- Alignment enum with options: Center, TopLeft, TopCenter, TopRight, CenterLeft, CenterRight, BottomLeft, BottomCenter, BottomRight
- Constructor to set layout parameters (fit, alignment, min/max X/Y coordinates)
- Methods to get runtime Fit and Alignment values for the Rive canvas
- Helper method to create a new Layout instance with modified parameters

This commit also exports the Layout class and enums from the rive.ts entry point.
