+++ js/src/layout/enums/Alignment.enum.ts
@@ -1,0 +1,41 @@
+/**
+ * Enum representing the different alignment options for the canvas.
+ */
+export enum Alignment {
+  /**
+   * Aligns the canvas content to the center.
+   */
+  Center = "center",
+  /**
+   * Aligns the canvas content to the top-left corner.
+   */
+  TopLeft = "topLeft",
+  /**
+   * Aligns the canvas content to the top center.
+   */
+  TopCenter = "topCenter",
+  /**
+   * Aligns the canvas content to the top-right corner.
+   */
+  TopRight = "topRight",
+  /**
+   * Aligns the canvas content to the center-left.
+   */
+  CenterLeft = "centerLeft",
+  /**
+   * Aligns the canvas content to the center-right.
+   */
+  CenterRight = "centerRight",
+  /**
+   * Aligns the canvas content to the bottom-left corner.
+   */
+  BottomLeft = "bottomLeft",
+  /**
+   * Aligns the canvas content to the bottom center.
+   */
+  BottomCenter = "bottomCenter",
+  /**
+   * Aligns the canvas content to the bottom-right corner.
+   */
+  BottomRight = "bottomRight",
+}
