+++ js/src/layout/enums/index.ts
@@ -1,0 +1,2 @@
+export { Alignment } from "./Alignment.enum";
+export { Fit } from "./Fit.enum";
