+++ js/src/layout/types/index.ts
@@ -1,0 +1,1 @@
+export type { LayoutParameters } from "./LayoutParameters.type";
